check-in content
- update check-in content modularization
- update check-in css styling with @mediq query
- update title
- update icon on tap
