+++ src/App.js
@@ -4,34 +4,8 @@ import React, { useState } from "react";
 import { useData } from './utilities/firebase.js';
 import Button from "./components/Button";
 import Banner from "./components/Banner";
+import Content from "./components/Content";
 
-
-
-
-const Content = ({ checkedIn, numPeople }) => {
-  // for purposes of demo:
-  // Low: 1 
-  // medium: 2
-  // high: 3
-
-  // actual expected values:
-  // low: <15
-  // medium: 15-40
-  // high: >40
-  return (
-    checkedIn ? 
-      <div>
-        <p>Jacobs Center Rapid Test - Covid-19</p>
-        <p>You're checked in! Don't forgot to check out!</p>
-      </div> :
-      <div>
-        <p>Jacobs Center Rapid Test - Covid-19</p>
-        <p>Low</p>
-        <p>Traffic</p>
-        <p>Approx. People in Line: {numPeople}</p>
-      </div>
-  )
-}
 
 function App() {
   const [checkedIn, setCheckedIn] = useState(false);
